+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.2] - 2020-04-29
+### Added
+* Storage: receive and handle data sent by the recorder
+* InstanceMethod: a single instance method on any class
+* ExampleCall: a single call to a given method and the parameter values that were passed
+* Added various development gem dependencies
+* Added rdoc documentation
+
 ## [0.1.1] - 2020-04-29
 ### Added
 * Github Actions for testing via rspec